requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
networkx>=3.3
python-hcl2>=4.3.4
pyyaml>=6.0.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.2.0
pydantic>=2.7.0
